Hotfix of longitudinal tracking for non-proton-type beams
The calculation of the synchronous phase takes the
charge given at the instantiation of RFSystems.
This argument is new and defaults to the proton value.

The setting of the charge only affects the phase locking
for multi-harmonic systems, otherwise phi_s is not used
internally.

This hotfix should correct for electron use as Xu Haisheng
(PSI) pointed out. Thanks for making the point.