pm: mean_per_slice, reorder in beam
- Dispatch for mean_per_slice, still a bit messy but works correctly
- Dispatch the reordering/sorting of the beam in particles to pm
Tests for the above work.
Problem with the per_slice functions: The CPU/GPU do not have the same
interface (see c++ and cobra functions require different parameters).
The solution was to add another layer of indirection which uses the same
interface for both: the sliceset and the array to operate on