BunchMonitor/SliceMonitor prep for GPU
This commit slightly changes the implementation of the BunchMonitor
and SliceMonitor such that it is easier to make use of GPUs (not
included here). The issue was the np.roll which is not available for
gpuarrays.
Solution: Use np.roll only when writing to file (data is on CPU anyway).
          Minimal changes to the implementation.
NO CHANGES TO THE INTERFACE

Details:
- write_to_buffer: do not np.roll array there but only in the write_to_file
  function. Fill the buffer by moving the write location.
- write_to_file: create an extra buffer, copy self.buffer into there and
  try to move it from the gpu. np.roll in this function, the rest stays
  the same.

- First unit tests for the slicemonitor and bunchmonitor