RFBucket: adapted the z_offset calculation.
Works much better for multi-harmonics to determine
the correct separatrix value. Before, the bucket
interval was not chosen correctly as the outer UFP
locations are influenced by the higher harmonics.

Tested with various settings below and above
transition, accelerating and decelerating.

The only problem with the separatrix was found for
above transition, decelerating with the following
values for RFSystems:

>>> print (rfsystems.voltages)
[5150.0, 7000.000000000001]
>>> print (rfsystems.harmonics)
[1, 2]
>>> print (rfsystems.phi_offsets)
[2.104717653589793, 5.601217653589793]
>>> print (rfsystems.alpha_array)
[0.5]
>>> print (rfsystems.p_increment)
-1e-24

As this is an unusual case, this may be fixed at a later point.