Hotfix: particles.p0 has been calculated slightly wrong.
Thanks to Mauro Migliorati for noticing the bug which
effectively lead to about 7% too large momentum increase
during acceleration (per turn). The problem was the formula
to calculate the gamma when changing p0 -- it involved
beta (but didn't calculate it new, instead it took the one
from before). Now it is self-consistent and correct.