pmath: put GPU version of arange into gpu_wrap
Like this it is available for functions inside gpu_wrap.
If functions are defined in the spill dictionaries in
pmath, they will only be available after gpu_wrap has been
imported. So we cannot use them in gpu_wrap.
Therefore all functionality used in gpu_wrap needs to be
defined there already, otherwise we end up with a circular
definition.