setup.py: making setup.py ready for pip install
exclude f2py stuff to Makefile which is not served to
the pip version. This way only one package gets created
and the UNKNOWN package due to the second setup call
from numpy.distutils vanishes. We do not need the errfff
(leading to this second separate setup call) any way
in the default TransverseGaussianSpaceCharge class.
By default, the scipy errfff is used.