synchrotron: allow higher-order chroma
Comparing Qp_x to 0 is a problem when providing
higher-order chroma in terms of a np.array in Qp_x.
The Synchrotron class therefore did not allow
higher-order chroma up until now if you provide it
as a np.array instead of a list. This commit fixes it,
thanks to Mario for spotting.