slicing: fixed bug in convert_to_particles
If empty_particles is not given, pmath.empty_like
was called with a dtype given which is not allowed.

Furthermore, now make sure that particle indices
outside the slicing area will be assigned zero.

Also, using the pmath.take now instead of the array
take which does not exist for GPUArrays. (Nonetheless,
this method fails for GPUArrays with an "invalid
subindex in axis 0" IndexError at the fancy
indexing assignment.)