spacecharge: added new FrozenGaussianSpaceCharge25D
The model precomputes a Gaussian electric field on a grid
via the TransverseGaussianSpaceCharge. This static field
is applied as kicks to the beam particles in a
FieldMapSliceWise fashtion, slice-by-slice,
with the weighted local line charge density to the beam.
That is to say, the longitudinal plane is self-consistent
while the transverse plane is a frozen Bassetti-Erskine
computed errf field of a fixed RMS beam size.
The interpolation is executed via PyPIClib's (PyPIC.GPU's)
field_to_particles (m2p) method.