pdf_integrators_2d: reviving and restructuring 2D quad integrators
Allow to compute 2D particle distribution function moments up
to second order for a domain parametrised by
[xmin, xmax], [ymin(x), ymax(x)].
This is particularly useful for RFBucket matching within a
RFBucket.interval for x and the separatrix functions for y.

Started with the dblquad based integrators, cumtrapz is to come.
So far the RFBucketMatchor still works nicely with these functions!