rf_bucket: total_potential may compute without shifting...
... to have zero potential energy at the separatrix.
The reason is that this reshifting of total_potential requires
already determined fix points. However, in some situations
(e.g. during the determination of the stable fix points) we
only need potential differences such that the absolute potential
value does not play a role -- here we can use the new
offset=False to avoid the reshifting.