pypic_factory: create_mesh now uses mesh.nz = nslices + 1
Since the foremost slice particles should be included
in the PyPIClib GPU algorithms (specifically the atomicAdd
variants) for space charge, we need the mesh to include one more
node so that the number of cells in between the nodes are equal
to the number of slices (fence post error).
This was wrong before which is why the particles in the front were
excluded from the particles_to_mesh method (which retrieves
the charge density on the mesh).