monitors: CellMonitor uses coordinates instead of bunch object
cobra_functions.stats.calc_cell_stats uses the explicit
coordinates now instead of the bunch object. CellMonitor in
monitors.py has correspondingly adapted and, in the case of
GPU contexts, moves the coordinates to the CPU before.

It would be much nicer to calculate the stats instead of in
cython on the CPU side directly on the GPU and only transfer
the GPU buffer (which is a CPU buffer right now) to the CPU.
This would speed up the CellMonitor by a lot!