Default value for circumference in slicer set to None.
We found an inconsistency: we have two circumferences, one in the synchrotron and one in the slicer, which are used in different ways. We set circumference to 0 in the slicer for now, which will always work for single bunch, but in the long run, for the sake of consistency, it should explicitly be provided as the real circumference.