GPU: importing pycuda.autoinit missing
if one initialises the GPU device manually and not via
importing pycuda.autoinit externally outside of PyHEADTAIL,
then the previous state here was failing as pycuda.autoinit
is used to retrieve the context -- but it has not been imported.

adding an explicit import for now, a better solution in the
future would be to explicitly use the context from the stack
of contexts in the PyCUDA.driver module.