GPU: removing automatic use of streams...
... and introducing flag use_streams in gpu_utils
to reenable in the future.

Streams are mostly used for concurrent launches of
kernels computing bunch statistics, both for
slicing as well as the monitors.

Removing the streams leads to serialising which slows
down marginally but makes the simulations more general.
(Here the reason is to interact with other libraries
like SixTrackLib which does not work in streamed computations)