SliceMonitor: Fixing stream.synchronize() calls on None
With the recent change to not use streams, they return None for the
functions calling the stream pool.
For the slice monitor, which calls gpu_wrap.sorted_emittances_per_slice()
this resulted in calling synchronize() on the streams which are None.

This patch fixes the None streams issue, thanks to Rob Williamson for the bug report.