bug fix in long space charge: the kicks omitted 1/dz behaviour
Thanks a lot to David Kelliher from STFC for pointing out this bug.
The kicks scaled with the number of slices since the division by
the slice widths was missing.
This bug is traceable to the restructuring of the SliceSet in 2015 at least.

Now the computed force (via make_force) and the actual tracked dp kicks
match.